+++ Example/SmileID/EnhancedKYC/EnhancedKycWithIdInputScreen.swift
@@ -47,18 +47,21 @@ struct EnhancedKycWithIdInputScreen: View {
                 productName: "ID",
                 partnerPrivacyPolicy: URL(string: "https://usesmileid.com")!,
                 showAttribution: true,
-                onConsentGranted: {
+                onConsentGranted: { consentInformation in
                     viewModel.onConsentGranted(
                         country: country,
                         idType: idType,
-                        requiredFields: requiredFields)
+                        consentInformation: consentInformation,
+                        requiredFields: requiredFields
+                    )
                 },
                 onConsentDenied: { delegate.didError(error: SmileIDError.consentDenied) }
             )
-        case .idInput(let country, let idType, let requiredFields):
+        case .idInput(let country, let idType, let consentInformation, let requiredFields):
             IdInfoInputScreen(
                 selectedCountry: country,
                 selectedIdType: idType,
+                consentInformation: consentInformation,
                 header: "Enter ID Information",
                 requiredFields: requiredFields,
                 onResult: viewModel.onIdFieldsEntered